changes to test added
